Add animation to texts
