+++ source/scss/blocks/parsed-text.scss
@@ -1,0 +1,24 @@
+.animation-screen {
+  position: fixed;
+  display: none;
+  justify-content: center;
+  align-items: center;
+  width: 100%;
+  height: 100%;
+  background-image: url("../img/scene.png");
+  background-repeat: no-repeat;
+  background-size: 100% auto;
+  background-position: center;
+
+  @include retina {
+    background-image: url("../img/scene@2x.png");
+  }
+
+  @media (orientation: portrait) {
+    background-image: url("../img/scene-mob.png");
+
+    @include retina {
+      background-image: url("../img/scene-mob@2x.png");
+    }
+  }
+}
