+++ source/scss/blocks/prizes.scss
@@ -330,3 +330,40 @@
     }
   }
 }
+
+.active {
+  .prizes__title {
+    .word {
+      display: inline-block;
+      overflow-y: hidden;
+      padding-top: 1.5rem;
+      line-height: 2.5rem;
+
+      .character {
+        display: inline-block;
+        animation: slide-in-bottom 0.8s ease-out both;
+
+        &:nth-child(3) {
+          animation-delay: 1s;
+          animation-duration: 0.8s;
+        }
+
+        &:nth-child(2),
+        &:nth-child(4) {
+          animation-delay: 1.1s;
+          animation-duration: 0.7s;
+        }
+
+        &:nth-child(5) {
+          animation-delay: 1.2s;
+          animation-duration: 0.6s;
+        }
+
+        &:nth-child(1) {
+          animation-delay: 1.3s;
+          animation-duration: 0.5s;
+        }
+      }
+    }
+  }
+}
