+++ source/scss/general/animations.scss
@@ -1,0 +1,9 @@
+@keyframes slide-in-bottom {
+  0% {
+    transform: translate3d(0, 170%, 0);
+  }
+
+  100% {
+    transform: translate3d(0, 0, 0);
+  }
+}
