+++ source/scss/style.scss
@@ -4,6 +4,7 @@
 @import "general/mixins.scss";
 @import "general/fonts.scss";
 @import "general/scaffolding.scss";
+@import "general/animations.scss";
 @import "blocks/animation-screen.scss";
 @import "blocks/btn.scss";
 @import "blocks/chat.scss";
